+++ src/components/index.js
@@ -1,3 +1,4 @@
+export { default as Container } from './Container';
 export { default as Button } from './Button';
 export { default as Card } from './Card';
 export { default as ModalContainer } from './Modal';